% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{annotation,NanoStringExperiment-method}
\alias{annotation,NanoStringExperiment-method}
\title{Access experiment annotations}
\usage{
\S4method{annotation}{NanoStringExperiment}(object)
}
\description{
Access annotations associated with the experiment
}
