% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-summary.R
\name{as.list,NanoStringExperiment-method}
\alias{as.list,NanoStringExperiment-method}
\title{Coerce NanoStringExperiment object to list}
\usage{
\S4method{as.list}{NanoStringExperiment}(x, ...)
}
\description{
Coercion method for NanoStringExperiment in list format
}
