% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{assayData,NanoStringExperiment-method}
\alias{assayData,NanoStringExperiment-method}
\title{Access assays}
\usage{
\S4method{assayData}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.
}
