% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{assayDataElement,NanoStringExperiment-method}
\alias{assayDataElement,NanoStringExperiment-method}
\title{Access assay data}
\usage{
\S4method{assayDataElement}{NanoStringExperiment}(object, elt)
}
\description{
It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayDataElement(testExp, "exprs")

}
