% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{assayDataElement}
\alias{assayDataElement}
\title{Access assay data}
\usage{
assayDataElement(object, elt)
}
\description{
It is recommended to use the SummarizedExperiment method assays instead.
This is a convenience method for backwards compatibility.
}
