% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{assayDataElementNames,NanoStringExperiment-method}
\alias{assayDataElementNames,NanoStringExperiment-method}
\title{Access assay names}
\usage{
\S4method{assayDataElementNames}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment assayNames instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayDataElementNames(testExp)

}
