% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{assayDataElementNames}
\alias{assayDataElementNames}
\title{Access assay names}
\usage{
assayDataElementNames(object)
}
\description{
It is recommended to use the SummarizedExperiment method assayNames instead.
This is a convenience method for backwards compatibility.
}
