% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{controlSubset}
\alias{controlSubset}
\title{Subset to control features}
\usage{
controlSubset(object)
}
\value{
NanoStringExperiment object
}
\description{
Method to subset to features with CodeClass IsControl
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
controlSubset(testExp)

}
