% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{design,NanoStringExperiment-method}
\alias{design,NanoStringExperiment-method}
\title{Access design formula}
\usage{
\S4method{design}{NanoStringExperiment}(object)
}
\description{
Access formula to be used for design in analyses
}
