% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{design<-,NanoStringExperiment,formula-method}
\alias{design<-,NanoStringExperiment,formula-method}
\title{Replace design formula}
\usage{
\S4method{design}{NanoStringExperiment,formula}(object) <- value
}
\description{
Replace or assign formula to be used for design in analyses
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
design(testExp) <- stats::as.formula(~ `cell_line`)

}
