% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{design}
\alias{design}
\title{Access design formula}
\usage{
design(object)
}
\description{
Access formula to be used for design in analyses
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
design(testExp)

}
