% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{dimLabels,NanoStringExperiment-method}
\alias{dimLabels,NanoStringExperiment-method}
\title{Access variables used for feature and sample identifiers}
\usage{
\S4method{dimLabels}{NanoStringExperiment}(object)
}
\description{
Shows variables used for rownames in rowData and colData.
}
