% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{dimLabels<-,NanoStringExperiment,ANY-method}
\alias{dimLabels<-,NanoStringExperiment,ANY-method}
\title{Replace variables used for feature and sample identifiers}
\usage{
\S4method{dimLabels}{NanoStringExperiment,ANY}(object) <- value

\S4method{dimLabels}{NanoStringExperiment,ANY}(object) <- value
}
\description{
Replaces rownames in rowData and colData with specified variable columns.

Replaces rownames in rowData and colData with specified variable columns.
}
