% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{dimLabels,NanoStringExperiment-method}
\alias{dimLabels,NanoStringExperiment-method}
\alias{dimLabels<-,NanoStringExperiment,ANY-method}
\title{Access variables used for feature and sample identifiers}
\usage{
\S4method{dimLabels}{NanoStringExperiment}(object)

\S4method{dimLabels}{NanoStringExperiment,ANY}(object) <- value
}
\value{
list of variables used for dimension labels

NanoStringExperiment object
}
\description{
Shows variables used for rownames in rowData and colData.

Replaces rownames in rowData and colData with specified variable columns.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
dimLabels(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
colData(testExp)[["testNames"]] <- paste0(dimnames(testExp)[[2]], "a")
dimLabels(testExp) <- c(dimLabels(testExp)[[1]], "testNames")

}
