% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-summary.R
\name{.marginalSummary}
\alias{.marginalSummary}
\title{Summary statistics list calculation}
\usage{
.marginalSummary(x, log2scale = TRUE)
}
\description{
Create summary statistics from list
}
