% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-summary.R
\docType{data}
\name{.summaryMetadata}
\alias{.summaryMetadata}
\title{Summary statistics list}
\format{
An object of class \code{list} of length 3.
}
\usage{
.summaryMetadata
}
\description{
Lists of useful summary statistics
}
\keyword{datasets}
