% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-apply.R
\name{esApply,NanoStringExperiment-method}
\alias{esApply,NanoStringExperiment-method}
\title{Apply across default exprs assay}
\usage{
\S4method{esApply}{NanoStringExperiment}(X, MARGIN, FUN, ...)
}
\description{
Apply function row (feature) or column (sample)-wise
to the default "exprs" assay
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
esApply(testExp, 1, mean)

}
