% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-apply.R
\name{esApply,NanoStringExperiment-method}
\alias{esApply,NanoStringExperiment-method}
\title{Apply across default exprs assay}
\usage{
\S4method{esApply}{NanoStringExperiment}(X, MARGIN, FUN, ...)
}
\value{
assay data matrix
}
\description{
Apply function row (feature) or column (sample)-wise
to the default "exprs" assay
}
\examples{
data(exampleNSEData)
esApply(testExp, 1, mean)

}
