% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{experimentData,NanoStringExperiment-method}
\alias{experimentData,NanoStringExperiment-method}
\title{Access experiment metadata}
\usage{
\S4method{experimentData}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment metadata instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
experimentData(testExp)

}
