% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{experimentData<-,NanoStringExperiment,ANY-method}
\alias{experimentData<-,NanoStringExperiment,ANY-method}
\title{Replace experiment metadata}
\usage{
\S4method{experimentData}{NanoStringExperiment,ANY}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment metadata instead.
This is a convenience method for backwards compatibility.
}
