% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{exprs<-,NanoStringExperiment,ANY-method}
\alias{exprs<-,NanoStringExperiment,ANY-method}
\title{Replace exprs data}
\usage{
\S4method{exprs}{NanoStringExperiment,ANY}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
exprs(testExp) <- exprs(testExp) + 1

}
