% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{exprs<-,NanoStringExperiment-method}
\alias{exprs<-,NanoStringExperiment-method}
\title{Replace exprs data}
\usage{
\S4method{exprs}{NanoStringExperiment}(object, ...) <- value
}
\description{
It is recommended to use the SummarizedExperiment method assays instead.
This is a convenience method for backwards compatibility.
}
