% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{featureData,NanoStringExperiment-method}
\alias{featureData,NanoStringExperiment-method}
\title{Access feature metadata}
\usage{
\S4method{featureData}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
featureData(testExp)

}
