% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{featureData<-,NanoStringExperiment,ANY-method}
\alias{featureData<-,NanoStringExperiment,ANY-method}
\title{Replace feature metadata}
\usage{
\S4method{featureData}{NanoStringExperiment,ANY}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
featureData(testExp) <- featureData(testExp)[, 1:2]

}
