% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{fData,NanoStringExperiment-method}
\alias{fData,NanoStringExperiment-method}
\alias{fData<-,NanoStringExperiment,ANY-method}
\alias{fvarLabels,NanoStringExperiment-method}
\alias{fvarLabels<-,NanoStringExperiment-method}
\alias{featureData,NanoStringExperiment-method}
\alias{featureData<-,NanoStringExperiment,ANY-method}
\alias{featureData<-,DataFrame,ANY-method}
\title{Access feature metadata}
\usage{
\S4method{fData}{NanoStringExperiment}(object)

\S4method{fData}{NanoStringExperiment,ANY}(object) <- value

\S4method{fvarLabels}{NanoStringExperiment}(object)

\S4method{fvarLabels}{NanoStringExperiment}(object) <- value

\S4method{featureData}{NanoStringExperiment}(object)

\S4method{featureData}{NanoStringExperiment,ANY}(object) <- value

\S4method{featureData}{DataFrame,ANY}(object) <- value
}
\value{
matrix of feature (column) metadata

NanoStringExperiment object

list of feature metadata variables

NanoStringExperiment object

matrix of feature (row) metadata

NanoStringExperiment object

DataFrame of feature metadata
}
\description{
It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
fData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
fData(testExp) <- rowData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
fvarLabels(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
fvarLabels(testExp) <- paste0(fvarLabels(testExp), "a")

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
featureData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
featureData(testExp) <- featureData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
featureData(testExp) <- featureData(testExp)[, 1:2]

}
