% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{featureNames,NanoStringExperiment-method}
\alias{featureNames,NanoStringExperiment-method}
\title{Access feature identifiers}
\usage{
\S4method{featureNames}{NanoStringExperiment}(object)
}
\description{
Shows variables used for rownames in rowData.
}
