% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{featureNames<-,NanoStringExperiment-method}
\alias{featureNames<-,NanoStringExperiment-method}
\title{Replace feature identifiers}
\usage{
\S4method{featureNames}{NanoStringExperiment}(object) <- value
}
\description{
Replace variables used for rownames in rowData.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
rowData(testExp)[["testNames"]] <- paste0(dimnames(testExp)[[1]], "a")
featureNames(testExp) <- rowData(testExp)[["testNames"]]

}
