% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{fvarLabels,NanoStringExperiment-method}
\alias{fvarLabels,NanoStringExperiment-method}
\title{Access feature metadata variables}
\usage{
\S4method{fvarLabels}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.
}
