% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{housekeepingSubset,NanoStringExperiment-method}
\alias{housekeepingSubset,NanoStringExperiment-method}
\title{Subset to housekeeping features}
\usage{
\S4method{housekeepingSubset}{NanoStringExperiment}(object)
}
\value{
NanoStringExperiment object
}
\description{
Method to subset to features with CodeClass Housekeeping
}
\examples{
data(exampleNSEData)
housekeepingSubset(testExp)

}
