% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{housekeepingSubset}
\alias{housekeepingSubset}
\title{Subset to housekeeping features}
\usage{
housekeepingSubset(object)
}
\value{
NanoStringExperiment object
}
\description{
Method to subset to features with CodeClass Housekeeping
}
\examples{
data(exampleNSEData)
housekeepingSubset(testExp)

}
