% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-munge.R
\name{munge,NanoStringExperiment-method}
\alias{munge,NanoStringExperiment-method}
\title{Append metadata to assay in long format}
\usage{
\S4method{munge}{NanoStringExperiment}(
  data,
  mapping = update(design(data), exprs ~ .),
  extradata = NULL,
  elt = "exprs",
  ...
)
}
\description{
Convert assay DataFrame into long format and append
labels and meta to the new DataFrame
}
