% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-class.R
\docType{class}
\name{NanoStringExperiment-class}
\alias{NanoStringExperiment-class}
\alias{.NanoStringExperiment}
\title{The NanoStringExperiment class}
\description{
Class definition for the base NanoString experiment
}
