% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{negativeControlSubset,NanoStringExperiment-method}
\alias{negativeControlSubset,NanoStringExperiment-method}
\title{Subset to negative control features}
\usage{
\S4method{negativeControlSubset}{NanoStringExperiment}(object)
}
\description{
Method to subset to features with CodeClass Negative
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
negativeControlSubset(testExp)

}
