% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{nonControlSubset,NanoStringExperiment-method}
\alias{nonControlSubset,NanoStringExperiment-method}
\title{Subset to non-IsControl features}
\usage{
\S4method{nonControlSubset}{NanoStringExperiment}(object)
}
\value{
NanoStringExperiment object
}
\description{
Method to subset to features not CodeClass IsControl
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
nonControlSubset(testExp)

}
