% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{nonControlSubset}
\alias{nonControlSubset}
\title{Subset to non-IsControl features}
\usage{
nonControlSubset(object)
}
\arguments{
\item{object}{NanoStringExperiment object}
}
\value{
NanoStringExperiment object
}
\description{
Method to subset to features not CodeClass IsControl
}
\examples{
data(exampleNSEData)
nonControlSubset(testExp)

}
