% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{pData,NanoStringExperiment-method}
\alias{pData,NanoStringExperiment-method}
\title{Access sample phenotypic metadata}
\usage{
\S4method{pData}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
