% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{pData<-,DataFrame,ANY-method}
\alias{pData<-,DataFrame,ANY-method}
\title{Replace sample phenotypic metadata}
\usage{
\S4method{pData}{DataFrame,ANY}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
