% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{pData<-,DataFrame-method}
\alias{pData<-,DataFrame-method}
\title{Replace sample phenotypic metadata}
\usage{
\S4method{pData}{DataFrame}(object, ...) <- value
}
\description{
It is recommended to use the SummarizedExperiment method colData instead.
This is a convenience method for backwards compatibility.
}
