% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{phenoData<-,DataFrame,ANY-method}
\alias{phenoData<-,DataFrame,ANY-method}
\title{Replace sample phenotypic metadata}
\usage{
\S4method{phenoData}{DataFrame,ANY}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
phenoData(testExp) <- phenoData(testExp)[, 1:2]

}
