% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{phenoData<-,NanoStringExperiment,ANY-method}
\alias{phenoData<-,NanoStringExperiment,ANY-method}
\title{Replace sample phenotypic metadata}
\usage{
\S4method{phenoData}{NanoStringExperiment,ANY}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment method colData instead.
This is a convenience method for backwards compatibility.
}
