% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{phenoData}
\alias{phenoData}
\title{Access sample phenotypic metadata}
\usage{
phenoData(object)
}
\description{
It is recommended to use the SummarizedExperiment method colData instead.
This is a convenience method for backwards compatibility.
}
