% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{positiveControlSubset,NanoStringExperiment-method}
\alias{positiveControlSubset,NanoStringExperiment-method}
\title{Subset to positive control features}
\usage{
\S4method{positiveControlSubset}{NanoStringExperiment}(object)
}
\arguments{
\item{object}{NanoStringExperiment object}
}
\value{
NanoStringExperiment object
}
\description{
Method to subset to features with CodeClass Positive
}
\examples{
data(exampleNSEData)
positiveControlSubset(testExp)

}
