% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{protocolData<-,DataFrame,ANY-method}
\alias{protocolData<-,DataFrame,ANY-method}
\title{Replace sample protocol metadata}
\usage{
\S4method{protocolData}{DataFrame,ANY}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment method colData instead.
This is a convenience method for backwards compatibility.
}
