% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{protocolData<-,NanoStringExperiment,ANY-method}
\alias{protocolData<-,NanoStringExperiment,ANY-method}
\title{Replace sample protocol metadata}
\usage{
\S4method{protocolData}{NanoStringExperiment,ANY}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
protocolData(testExp) <- protocolData(testExp)[, 1:2]

}
