% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{protocolData}
\alias{protocolData}
\title{Access sample protocol metadata}
\usage{
protocolData(object)
}
\description{
It is recommended to use the SummarizedExperiment method colData instead.
This is a convenience method for backwards compatibility.
}
