% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{sData}
\alias{sData}
\alias{sData,NanoStringExperiment-method}
\alias{svarLabels}
\alias{svarLabels,NanoStringExperiment-method}
\alias{phenoData,NanoStringExperiment-method}
\alias{phenoData<-,NanoStringExperiment,ANY-method}
\alias{protocolData,NanoStringExperiment-method}
\alias{protocolData<-,NanoStringExperiment,ANY-method}
\alias{pData,NanoStringExperiment-method}
\alias{pData,DataFrame-method}
\alias{pData<-,NanoStringExperiment,ANY-method}
\alias{varLabels,NanoStringExperiment-method}
\alias{varLabels,DataFrame-method}
\alias{varLabels<-,NanoStringExperiment-method}
\title{Access sample metadata}
\usage{
sData(object)

\S4method{sData}{NanoStringExperiment}(object)

svarLabels(object)

\S4method{svarLabels}{NanoStringExperiment}(object)

\S4method{phenoData}{NanoStringExperiment}(object)

\S4method{phenoData}{NanoStringExperiment,ANY}(object) <- value

\S4method{protocolData}{NanoStringExperiment}(object)

\S4method{protocolData}{NanoStringExperiment,ANY}(object) <- value

\S4method{pData}{NanoStringExperiment}(object)

\S4method{pData}{DataFrame}(object)

\S4method{pData}{NanoStringExperiment,ANY}(object) <- value

\S4method{varLabels}{NanoStringExperiment}(object)

\S4method{varLabels}{DataFrame}(object)

\S4method{varLabels}{NanoStringExperiment}(object) <- value
}
\arguments{
\item{object}{NanoStringExperiment object}

\item{value}{replacement variable labels}
}
\value{
matrix of sample (column) metadata

matrix of sample (column) metadata

list of sample (column) metadata variables

list of sample (column) metadata variables

matrix of sample phenotypic metadata

NanoStringExperiment object

matrix of sample protocol metadata

NanoStringExperiment object

matrix of sample phenotypic metadata

DataFrame of sample phenotypic metadata

NanoStringExperiment object

list of sample phenotypic metadata variables

list of metadata variables

NanoStringExperiment object
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
\examples{
data(exampleNSEData)
sData(testExp)

data(exampleNSEData)
sData(testExp)

data(exampleNSEData)
svarLabels(testExp)

data(exampleNSEData)
svarLabels(testExp)

data(exampleNSEData)
phenoData(testExp)

data(exampleNSEData)
phenoData(testExp) <- phenoData(testExp)[, 1:2]

data(exampleNSEData)
protocolData(testExp)

data(exampleNSEData)
protocolData(testExp) <- protocolData(testExp)[, 1:2]

data(exampleNSEData)
pData(testExp)

data(exampleNSEData)
pData(testExp)

data(exampleNSEData)
pData(testExp) <- pData(testExp)[, 1:2]

data(exampleNSEData)
varLabels(testExp)

data(exampleNSEData)
varLabels(phenoData(testExp))

data(exampleNSEData)
varLabels(testExp) <- paste0(varLabels(testExp), "a")

}
