% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{sampleNames,NanoStringExperiment-method}
\alias{sampleNames,NanoStringExperiment-method}
\title{Access sample identifiers}
\usage{
\S4method{sampleNames}{NanoStringExperiment}(object)
}
\description{
Shows variables used for rownames in colData.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
sampleNames(testExp)

}
