% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{sampleNames<-,NanoStringExperiment,ANY-method}
\alias{sampleNames<-,NanoStringExperiment,ANY-method}
\title{Replace sample identifiers}
\usage{
\S4method{sampleNames}{NanoStringExperiment,ANY}(object) <- value
}
\description{
Replace variables used for rownames in colData.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
colData(testExp)[["testNames"]] <- paste0(dimnames(testExp)[[2]], "a")
sampleNames(testExp) <- colData(testExp)[["testNames"]]

}
