% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{sData,NanoStringExperiment-method}
\alias{sData,NanoStringExperiment-method}
\title{Access sample metadata}
\usage{
\S4method{sData}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
