% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-qc.R
\name{setQCFlags}
\alias{setQCFlags}
\title{Set quality control flags}
\usage{
setQCFlags(object, ...)
}
\arguments{
\item{object}{NanoStringExperiment object}

\item{...}{additional parameters to pass to methods}
}
\value{
NanoStringExperiment object
}
\description{
Flag samples or features that fail quality control checks
}
