% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{signatures,NanoStringExperiment-method}
\alias{signatures,NanoStringExperiment-method}
\title{Access signatures slot}
\usage{
\S4method{signatures}{NanoStringExperiment}(object)
}
\value{
SignatureSet object
}
\description{
Access signatures slot to get \code{SignatureSet} object
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
signatures(testExp)

}
