% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureSet-class.R
\docType{class}
\name{SignatureSet-class}
\alias{SignatureSet-class}
\alias{.SignatureSet}
\alias{initialize,SignatureSet-method}
\alias{show,SignatureSet-method}
\alias{SignatureSet}
\title{The SignatureSet class}
\usage{
\S4method{initialize}{SignatureSet}(
  .Object,
  weights = NumericList(),
  groups = factor(),
  func = character(),
  version = character(),
  ...
)

\S4method{initialize}{SignatureSet}(
  .Object,
  weights = NumericList(),
  groups = factor(),
  func = character(),
  version = character(),
  ...
)

\S4method{initialize}{SignatureSet}(
  .Object,
  weights = NumericList(),
  groups = factor(),
  func = character(),
  version = character(),
  ...
)

\S4method{initialize}{SignatureSet}(
  .Object,
  weights = NumericList(),
  groups = factor(),
  func = character(),
  version = character(),
  ...
)

\S4method{show}{SignatureSet}(object)

SignatureSet(
  weights = NumericList(),
  groups = factor(),
  func = character(),
  version = character(),
  ...
)

SignatureSet(
  weights = NumericList(),
  groups = factor(),
  func = character(),
  version = character(),
  ...
)

SignatureSet(
  weights = NumericList(),
  groups = factor(),
  func = character(),
  version = character(),
  ...
)

SignatureSet(
  weights = NumericList(),
  groups = factor(),
  func = character(),
  version = character(),
  ...
)
}
\value{
SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object

summary of SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object
}
\description{
Class definition for a SignatureSet. Holds weights and grouping for 
expression signature groups.

Initialize SignatureSet with weights

Initialize SignatureSet with weights and groups

Initialize SignatureSet with weights, groups, and func

Initialize SignatureSet with weights, groups, func, and version

Show method for SignatureSet

Generate new SignatureSet with weights

Generate new SignatureSet with weights and groups

Generate new SignatureSet with weights, groups, and func

Generate new SignatureSet with weights, groups, func, and version
}
\section{Slots}{

\describe{
\item{\code{\code{weights}}}{named \code{NumericList} defining signatures based 
on linear combinations of genes.}

\item{\code{\code{groups}}}{factor vector indicating groups in \code{SignatureSet}}

\item{\code{\code{func}}}{character indicating function to use}

\item{\code{\code{version}}}{character indicating version to use}
}}

\examples{
testSig <- SignatureSet()
show(testSig)

}
