% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureSet-methods.R
\name{weights,SignatureSet-method}
\alias{weights,SignatureSet-method}
\alias{weights<-}
\alias{weights<-,SignatureSet,NumericList-method}
\alias{weights<-,SignatureSet,CompressedNumericList-method}
\alias{weights<-,SignatureSet,ANY-method}
\alias{weights<-,SignatureSet,list-method}
\alias{weights<-,SignatureSet,NULL-method}
\alias{groups}
\alias{groups,SignatureSet-method}
\alias{groups<-}
\alias{groups<-,SignatureSet,factor-method}
\alias{groups<-,SignatureSet,ANY-method}
\alias{groups<-,SignatureSet,NULL-method}
\alias{getSigFuncs}
\alias{getSigFuncs,SignatureSet-method}
\alias{setSigFuncs<-}
\alias{setSigFuncs<-,SignatureSet,character-method}
\alias{version}
\alias{version,SignatureSet-method}
\alias{version<-}
\alias{version<-,SignatureSet,character-method}
\alias{version<-,SignatureSet,ANY-method}
\alias{version<-,SignatureSet,NULL-method}
\alias{length,SignatureSet-method}
\alias{names,SignatureSet-method}
\alias{lengths,SignatureSet-method}
\title{Access weights}
\usage{
\S4method{weights}{SignatureSet}(object, ...)

weights(object) <- value

\S4method{weights}{SignatureSet,NumericList}(object) <- value

\S4method{weights}{SignatureSet,CompressedNumericList}(object) <- value

\S4method{weights}{SignatureSet,ANY}(object) <- value

\S4method{weights}{SignatureSet,list}(object) <- value

\S4method{weights}{SignatureSet,`NULL`}(object) <- value

groups(object, ...)

\S4method{groups}{SignatureSet}(object, ...)

groups(object) <- value

\S4method{groups}{SignatureSet,factor}(object) <- value

\S4method{groups}{SignatureSet,ANY}(object) <- value

\S4method{groups}{SignatureSet,`NULL`}(object) <- value

getSigFuncs(object, ...)

\S4method{getSigFuncs}{SignatureSet}(object, ...)

setSigFuncs(object) <- value

\S4method{setSigFuncs}{SignatureSet,character}(object) <- value

version(object, ...)

\S4method{version}{SignatureSet}(object, ...)

version(object) <- value

\S4method{version}{SignatureSet,character}(object) <- value

\S4method{version}{SignatureSet,ANY}(object) <- value

\S4method{version}{SignatureSet,`NULL`}(object) <- value

\S4method{length}{SignatureSet}(x)

\S4method{names}{SignatureSet}(x)

\S4method{lengths}{SignatureSet}(x, use.names = TRUE)
}
\value{
NumericList of weights

SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object

factor vector

factor vector

SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object

character vector of weight functions

character vector of weight functions

SignatureSet object

SignatureSet object

character string indicating signature versions used

character string indicating signature versions used

SignatureSet object

SignatureSet object

SignatureSet object

SignatureSet object

numeric length

character vector

character vector
}
\description{
Access weights in weights slot

Replace or assign weights in weights slot

Replace or assign weights in weights slot

Replace or assign weights in weights slot

Replace or assign weights in weights slot

Replace or assign weights in weights slot

Replace or assign weights in weights slot

Access groups factor vector for SignatureSet object

Access groups factor vector for SignatureSet object

Replace or assign factor vector in groups slot

Replace or assign factor vector in groups slot

Replace or assign factor vector in groups slot

Replace or assign factor vector in groups slot

Access signature functions for weights

Access signature functions for weights

Replace signature function for SignatureSet object

Replace signature function for SignatureSet object

Access signature versions for SignatureSet object

Access signature versions for SignatureSet object

Replace signature versions for SignatureSet object

Replace signature versions for SignatureSet object

Replace signature versions for SignatureSet object

Replace signature versions for SignatureSet object

Get the length of weights NumericList corresponding
to the length of signatures

Get names of the weights NumericList corresponding
to the names of signatures

Get number of features within each weight element in the 
weights NumericList corresponding to the number of 
features in each signature.
}
