% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{signatureSubset,NanoStringExperiment-method}
\alias{signatureSubset,NanoStringExperiment-method}
\title{Subset to signature features}
\usage{
\S4method{signatureSubset}{NanoStringExperiment}(object)
}
\value{
NanoStringExperiment object
}
\description{
Method to subset to features in signatures slot
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
signatureSubset(testExp)

}
