% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{signatureSubset}
\alias{signatureSubset}
\title{Subset to signature features}
\usage{
signatureSubset(object)
}
\value{
NanoStringExperiment object
}
\description{
Method to subset to features in signatures slot
}
\examples{
data(exampleNSEData)
signatureSubset(testExp)

}
