% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-summary.R
\name{summary,NanoStringExperiment-method}
\alias{summary,NanoStringExperiment-method}
\title{Summarize assay data}
\usage{
\S4method{summary}{NanoStringExperiment}(
  object,
  MARGIN = 2L,
  GROUP = NULL,
  log2scale = TRUE,
  elt = "exprs",
  signatureScores = FALSE,
  ...
)
}
\arguments{
\item{object}{NanoStringExperiment object}

\item{MARGIN}{integer to apply across row or column}

\item{GROUP}{colData header to group by}

\item{log2scale}{boolean indicating to use log2 values}

\item{elt}{assay data matrix name}

\item{signatureScores}{boolean to use signature scores}

\item{...}{parameters to pass to summary}
}
\value{
summary table of aggregated results
}
\description{
Get chosen summary statistics for a selected assay
}
\examples{
data(exampleNSEData)
summary(testExp)

}
