% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{svarLabels}
\alias{svarLabels}
\title{Access sample metadata varible names}
\usage{
svarLabels(object)
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
