% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testPhenoData}
\alias{testPhenoData}
\title{NanoString sample phenotypic metadata}
\source{
NanoString Technologies
}
\description{
AnnotatedDataFrame of NanoString GeoMx phenotypic sample metadata
}
