% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testProtocolData}
\alias{testProtocolData}
\title{NanoString sample protocol metadata}
\source{
NanoString Technologies
}
\description{
AnnotatedDataFrame of NanoString GeoMx protocol sample metadata
}
