% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{varLabels<-,NanoStringExperiment-method}
\alias{varLabels<-,NanoStringExperiment-method}
\title{Replace sample phenotypic metadata variables}
\usage{
\S4method{varLabels}{NanoStringExperiment}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
