% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-summary.R
\name{with,NanoStringExperiment-method}
\alias{with,NanoStringExperiment-method}
\title{Evaluate expression on NanoStringExperiment list}
\usage{
\S4method{with}{NanoStringExperiment}(data, expr, ...)
}
\value{
results of evaluation
}
\description{
Performs expression within NanoStringExperiment list
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
with(testExp, table(CodeClass))

}
