% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{assayData,NanoStringExperiment-method}
\alias{assayData,NanoStringExperiment-method}
\alias{assayDataElement}
\alias{assayDataElement,NanoStringExperiment-method}
\alias{assayDataElement<-}
\alias{assayDataElement<-,NanoStringExperiment-method}
\alias{assayDataElementNames}
\alias{assayDataElementNames,NanoStringExperiment-method}
\alias{exprs,NanoStringExperiment-method}
\alias{exprs<-,NanoStringExperiment,ANY-method}
\title{Access assays}
\usage{
\S4method{assayData}{NanoStringExperiment}(object)

assayDataElement(object, elt)

\S4method{assayDataElement}{NanoStringExperiment}(object, elt)

assayDataElement(object, elt, ...) <- value

\S4method{assayDataElement}{NanoStringExperiment}(object, elt, ...) <- value

assayDataElementNames(object)

\S4method{assayDataElementNames}{NanoStringExperiment}(object)

\S4method{exprs}{NanoStringExperiment}(object)

\S4method{exprs}{NanoStringExperiment,ANY}(object) <- value
}
\value{
summary of assays in object

expression matrix for assay element

expression matrix for assay element

NanoStringExperiment object

NanoStringExperiment object

list of assay elements in object

list of assay elements in object

expression matrix for assay element \code{"exprs"}
}
\description{
It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment assayNames instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment assayNames instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayDataElement(testExp, "exprs")

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayDataElement(testExp, "exprs")

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayDataElement(testExp, "exprsShift") <- 
    assayDataElement(testExp, "exprs") + 1

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayDataElement(testExp, "exprsShift") <- 
    assayDataElement(testExp, "exprs") + 1

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayDataElementNames(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
assayDataElementNames(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
exprs(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
exprs(testExp) <- exprs(testExp) + 1

}
