% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{assayData}
\alias{assayData}
\title{Access assays}
\usage{
assayData(object)
}
\description{
It is recommended to use the SummarizedExperiment method assays instead.
This is a convenience method for backwards compatibility.
}
