% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-apply.R
\name{assayDataApply}
\alias{assayDataApply}
\title{Apply across an assay}
\usage{
assayDataApply(X, MARGIN, FUN, ..., elt = "exprs")

assayDataApply(X, MARGIN, FUN, ..., elt = "exprs")
}
\description{
Apply function row (feature) or column (sample)-wise
to selected assay

Apply function row (feature) or column (sample)-wise
to selected assay
}
