% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-apply.R
\name{assayDataApply}
\alias{assayDataApply}
\title{Apply across an assay}
\usage{
assayDataApply(X, MARGIN, FUN, ..., elt = "exprs")

assayDataApply(X, MARGIN, FUN, ..., elt = "exprs")
}
\arguments{
\item{X}{NanoStringExperiment object}

\item{MARGIN}{integer to apply across row or column}

\item{FUN}{function to apply}

\item{...}{parameters to pass to FUN}

\item{elt}{expression matrix for assay element \code{"exprs"}}
}
\value{
assay data matrix

assay data matrix
}
\description{
Apply function row (feature) or column (sample)-wise
to selected assay

Apply function row (feature) or column (sample)-wise
to selected assay
}
\examples{
data(exampleNSEData)
assayDataApply(testExp, 1, mean)

data(exampleNSEData)
assayDataApply(testExp, 1, mean, elt="exprs")

}
