% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{assayDataElement<-}
\alias{assayDataElement<-}
\title{Replace or add new assay data}
\usage{
assayDataElement(object, elt, ...) <- value
}
\description{
It is recommended to use the SummarizedExperiment assays instead.
This is a convenience method for backwards compatibility.
}
