% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{design<-,NanoStringExperiment,ANY-method}
\alias{design<-,NanoStringExperiment,ANY-method}
\title{Replace design formula}
\usage{
\S4method{design}{NanoStringExperiment,ANY}(object) <- value
}
\description{
Replace or assign formula to be used for design in analyses
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
design(testExp) <- ~ `cell_line`

}
