% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{dimLabels,NanoStringExperiment-method}
\alias{dimLabels,NanoStringExperiment-method}
\alias{dimLabels<-,NanoStringExperiment,ANY-method}
\title{Access variables used for feature and sample identifiers}
\usage{
\S4method{dimLabels}{NanoStringExperiment}(object)

\S4method{dimLabels}{NanoStringExperiment,ANY}(object) <- value
}
\arguments{
\item{object}{NanoStringExperiment object}

\item{value}{list of key headers to use for sample and feature names}
}
\value{
list of variables used for dimension labels

NanoStringExperiment object
}
\description{
Shows variables used for rownames in rowData and colData.

Replaces rownames in rowData and colData with specified variable columns.
}
\examples{
data(exampleNSEData)
dimLabels(testExp)

data(exampleNSEData)
colData(testExp)[["testNames"]] <- paste0(dimnames(testExp)[[2]], "a")
dimLabels(testExp) <- c(dimLabels(testExp)[[1]], "testNames")

}
