% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{endogenousSubset,NanoStringExperiment-method}
\alias{endogenousSubset,NanoStringExperiment-method}
\title{Subset to endogenous features}
\usage{
\S4method{endogenousSubset}{NanoStringExperiment}(object)
}
\description{
Method to subset to features with CodeClass Endogenous
}
