% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-apply.R
\name{esBy,NanoStringExperiment-method}
\alias{esBy,NanoStringExperiment-method}
\title{Apply across NanoStringExperiment object by groupings}
\usage{
\S4method{esBy}{NanoStringExperiment}(X, GROUP, FUN, ..., simplify = TRUE)
}
\description{
Group samples or features by variable and apply a function
}
\details{
esBy(testExp, 
    GROUP = "cell_line", 
    FUN = function(x) { 
        assayDataApply(x, MARGIN = 1, FUN = mean, elt = "exprs") 
    })
}
