% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{experimentData,NanoStringExperiment-method}
\alias{experimentData,NanoStringExperiment-method}
\alias{experimentData<-,NanoStringExperiment,ANY-method}
\alias{expinfo,list-method}
\alias{otherInfo,list-method}
\title{Access experiment metadata}
\usage{
\S4method{experimentData}{NanoStringExperiment}(object)

\S4method{experimentData}{NanoStringExperiment,ANY}(object) <- value

\S4method{expinfo}{list}(object)

\S4method{otherInfo}{list}(object)
}
\arguments{
\item{object}{list of experiment data}

\item{value}{list of experiment parameters}
}
\value{
list of experiment metadata

NanoStringExperiment object

list of experiment metadata

list of experiment metadata
}
\description{
It is recommended to use the SummarizedExperiment metadata instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment metadata instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment metadata instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment metadata instead.
This is a convenience method for backwards compatibility.
}
\examples{
data(exampleNSEData)
experimentData(testExp)

data(exampleNSEData)
experimentData(testExp) <- 
    c(experimentData(testExp), list(testName="myExp"))

data(exampleNSEData)
expinfo(experimentData(testExp))

data(exampleNSEData)
otherInfo(experimentData(testExp))

}
