% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{fData,NanoStringExperiment-method}
\alias{fData,NanoStringExperiment-method}
\title{Access feature metadata}
\usage{
\S4method{fData}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.
}
