% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{fData,NanoStringExperiment-method}
\alias{fData,NanoStringExperiment-method}
\alias{fData<-,NanoStringExperiment,ANY-method}
\alias{fvarLabels,NanoStringExperiment-method}
\alias{fvarLabels<-,NanoStringExperiment-method}
\alias{featureData,NanoStringExperiment-method}
\alias{featureData<-,NanoStringExperiment,ANY-method}
\title{Access feature metadata}
\usage{
\S4method{fData}{NanoStringExperiment}(object)

\S4method{fData}{NanoStringExperiment,ANY}(object) <- value

\S4method{fvarLabels}{NanoStringExperiment}(object)

\S4method{fvarLabels}{NanoStringExperiment}(object) <- value

\S4method{featureData}{NanoStringExperiment}(object)

\S4method{featureData}{NanoStringExperiment,ANY}(object) <- value
}
\arguments{
\item{object}{NanoStringExperiment object}

\item{value}{replacement feature metadata}
}
\value{
matrix of feature (column) metadata

NanoStringExperiment object

list of feature metadata variables

NanoStringExperiment object

matrix of feature (row) metadata

NanoStringExperiment object
}
\description{
It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.
}
\examples{
data(exampleNSEData)
fData(testExp)

data(exampleNSEData)
fData(testExp) <- rowData(testExp)[, 1:2]

data(exampleNSEData)
fvarLabels(testExp)

data(exampleNSEData)
fvarLabels(testExp) <- paste0(fvarLabels(testExp), "a")

data(exampleNSEData)
featureData(testExp)

data(exampleNSEData)
featureData(testExp) <- featureData(testExp)[, 1:2]

}
