% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{featureNames,NanoStringExperiment-method}
\alias{featureNames,NanoStringExperiment-method}
\alias{featureNames<-,NanoStringExperiment-method}
\title{Access feature identifiers}
\usage{
\S4method{featureNames}{NanoStringExperiment}(object)

\S4method{featureNames}{NanoStringExperiment}(object) <- value
}
\value{
list of string feature (row) identifiers

NanoStringExperiment object
}
\description{
Shows identifiers used for features (rownames) in assay

Replace variables used for rownames in rowData.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
featureNames(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
rowData(testExp)[["testNames"]] <- paste0(dimnames(testExp)[[1]], "a")
featureNames(testExp) <- rowData(testExp)[["testNames"]]

}
