% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{fvarLabels<-,NanoStringExperiment-method}
\alias{fvarLabels<-,NanoStringExperiment-method}
\title{Replace feature metadata variables}
\usage{
\S4method{fvarLabels}{NanoStringExperiment}(object) <- value
}
\description{
It is recommended to use the SummarizedExperiment rowData instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
fvarLabels(testExp) <- paste0(fvarLabels(testExp), "a")

}
