% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-munge.R
\name{munge}
\alias{munge}
\title{Append metadata to assay in long format}
\usage{
munge(data, ...)
}
\value{
matrix of results in long format
}
\description{
Convert assay DataFrame into long format and append
labels and meta to the new DataFrame
}
\examples{
data(exampleNSEData)
head(munge(testExp, mapping = ~`cell_line` + exprs))

}
