% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-class.R
\docType{class}
\name{NanoStringExperiment-class}
\alias{NanoStringExperiment-class}
\alias{.NanoStringExperiment}
\alias{show,NanoStringExperiment-method}
\alias{NanoStringExperiment}
\alias{NanoStringExperiment,missing-method}
\alias{NanoStringExperiment,matrix-method}
\title{The NanoStringExperiment class}
\usage{
\S4method{show}{NanoStringExperiment}(object)

NanoStringExperiment(
  assayData,
  phenoData = Biobase::annotatedDataFrameFrom(assayData, byrow = FALSE),
  protocolData = Biobase::annotatedDataFrameFrom(assayData, byrow = FALSE),
  featureData = Biobase::annotatedDataFrameFrom(assayData, byrow = TRUE),
  experimentData = Biobase::MIAME(),
  annotation = character(),
  dimLabels = character(),
  signatures = NanoStringNCTools::SignatureSet(),
  design = NULL,
  ...
)

\S4method{NanoStringExperiment}{missing}(
  assayData,
  phenoData = Biobase::annotatedDataFrameFrom(assayData, byrow = FALSE),
  protocolData = Biobase::annotatedDataFrameFrom(assayData, byrow = FALSE),
  featureData = Biobase::annotatedDataFrameFrom(assayData, byrow = TRUE),
  experimentData = Biobase::MIAME(),
  annotation = character(),
  dimLabels = character(),
  signatures = NanoStringNCTools::SignatureSet(),
  design = NULL,
  ...
)

\S4method{NanoStringExperiment}{matrix}(
  assayData,
  phenoData = Biobase::annotatedDataFrameFrom(assayData, byrow = FALSE),
  protocolData = Biobase::annotatedDataFrameFrom(assayData, byrow = FALSE),
  featureData = Biobase::annotatedDataFrameFrom(assayData, byrow = TRUE),
  experimentData = Biobase::MIAME(),
  annotation = character(),
  dimLabels = character(),
  signatures = NanoStringNCTools::SignatureSet(),
  design = NULL,
  ...
)
}
\arguments{
\item{object}{NanoStringExperiment object}

\item{assayData}{matrix of expression values}

\item{phenoData}{AnnotatedDataFrame with phenotype metadata}

\item{protocolData}{AnnotatedDataFrame with protocol metadata}

\item{featureData}{AnnotatedDataFrame with feature metadata}

\item{experimentData}{MIAME object}

\item{annotation}{list of annotations used}

\item{dimLabels}{list of key headers for row and column metadata}

\item{signatures}{SignatureSet object}

\item{design}{formula or NULL}

\item{...}{parameters to pass}
}
\value{
NanoStringExperiment object

summary of NanoStringExperiment object

NanoStringExperiment object

NanoStringExperiment object

NanoStringExperiment object
}
\description{
Class definition for the base NanoString experiment

Show method for NanoStringExperiment

Constructs NanoStringExperiment object based on assay data signature

Constructs NanoStringExperiment object if assay data missing

Constructs NanoStringExperiment object if assay data is a matrix
}
\section{Slots}{

\describe{
\item{\code{.__classVersion__}}{version of class}

\item{\code{assayData}}{expression matrix or NULL}

\item{\code{annotation}}{character list of annotations used}

\item{\code{dimLabels}}{character list of dimension labels}

\item{\code{signatures}}{optional SignatureSet object}

\item{\code{design}}{formula or NULL}
}}

\examples{
data(exampleNSEData)
show(testExp)

data(exampleNSEData)
NanoStringExperiment(
    assayData = testAssayData,
    phenoData = testPhenoData,
    protocolData = testProtocolData,
    featureData = testFeatureData,
    experimentData = testExperimentData,
    annotation = testAnnotation,
    dimLabels = testDimLabels)

NanoStringExperiment()

data(exampleNSEData)
NanoStringExperiment(
    assayData = testAssayData,
    phenoData = testPhenoData,
    protocolData = testProtocolData,
    featureData = testFeatureData,
    experimentData = testExperimentData,
    annotation = testAnnotation,
    dimLabels = testDimLabels)

}
