% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{negativeControlSubset}
\alias{negativeControlSubset}
\title{Subset to negative control features}
\usage{
negativeControlSubset(object)
}
\description{
Method to subset to features with CodeClass Negative
}
