% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize an assay}
\usage{
normalize(object, ...)
}
\arguments{
\item{object}{NanoStringExperiment object}

\item{...}{additional parameters to pass to methods}
}
\value{
NanoStringExperiment object
}
\description{
Normalize assay data based on chosen method
}
