% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-subset.R
\name{positiveControlSubset}
\alias{positiveControlSubset}
\title{Subset to positive control features}
\usage{
positiveControlSubset(object)
}
\description{
Method to subset to features with CodeClass Positive
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
positiveControlSubset(testExp)

}
