% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{sData}
\alias{sData}
\alias{sData,NanoStringExperiment-method}
\alias{svarLabels}
\alias{svarLabels,NanoStringExperiment-method}
\alias{phenoData,NanoStringExperiment-method}
\alias{phenoData<-,NanoStringExperiment,ANY-method}
\alias{phenoData<-,DataFrame,ANY-method}
\alias{phenoData<-,vector,ANY-method}
\alias{protocolData,NanoStringExperiment-method}
\alias{protocolData<-,NanoStringExperiment,ANY-method}
\alias{protocolData<-,DataFrame,ANY-method}
\alias{protocolData<-,vector,ANY-method}
\alias{pData,NanoStringExperiment-method}
\alias{pData,DataFrame-method}
\alias{pData<-,NanoStringExperiment,ANY-method}
\alias{pData<-,DataFrame,ANY-method}
\alias{pData<-,vector,ANY-method}
\alias{varLabels,NanoStringExperiment-method}
\alias{varLabels,DataFrame-method}
\alias{varLabels<-,NanoStringExperiment-method}
\alias{varLabels<-,vector-method}
\title{Access sample metadata}
\usage{
sData(object)

\S4method{sData}{NanoStringExperiment}(object)

svarLabels(object)

\S4method{svarLabels}{NanoStringExperiment}(object)

\S4method{phenoData}{NanoStringExperiment}(object)

\S4method{phenoData}{NanoStringExperiment,ANY}(object) <- value

\S4method{phenoData}{DataFrame,ANY}(object) <- value

\S4method{phenoData}{vector,ANY}(object) <- value

\S4method{protocolData}{NanoStringExperiment}(object)

\S4method{protocolData}{NanoStringExperiment,ANY}(object) <- value

\S4method{protocolData}{DataFrame,ANY}(object) <- value

\S4method{protocolData}{vector,ANY}(object) <- value

\S4method{pData}{NanoStringExperiment}(object)

\S4method{pData}{DataFrame}(object)

\S4method{pData}{NanoStringExperiment,ANY}(object) <- value

\S4method{pData}{DataFrame,ANY}(object) <- value

\S4method{pData}{vector,ANY}(object) <- value

\S4method{varLabels}{NanoStringExperiment}(object)

\S4method{varLabels}{DataFrame}(object)

\S4method{varLabels}{NanoStringExperiment}(object) <- value

\S4method{varLabels}{vector}(object) <- value
}
\value{
matrix of sample (column) metadata

matrix of sample (column) metadata

list of sample (column) metadata variables

list of sample (column) metadata variables

matrix of sample phenotypic metadata

NanoStringExperiment object

DataFrame of sample phenotypic metadata

DataFrame of sample phenotypic metadata

matrix of sample protocol metadata

NanoStringExperiment object

DataFrame of sample protocol metadata

DataFrame of sample protocol metadata

matrix of sample phenotypic metadata

DataFrame of sample phenotypic metadata

NanoStringExperiment object

DataFrame of sample phenotypic metadata

DataFrame of sample phenotypic metadata

list of sample phenotypic metadata variables

list of metadata variables

NanoStringExperiment object

DataFrame with updated column names
}
\description{
It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.

It is recommended to use the SummarizedExperiment colData instead.
This is a convenience method for backwards compatibility.
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
sData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
sData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
svarLabels(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
svarLabels(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
phenoData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
phenoData(testExp) <- phenoData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
phenoData(testExp) <- phenoData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
phenoData(testExp) <- phenoData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
protocolData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
protocolData(testExp) <- protocolData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
protocolData(testExp) <- protocolData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
protocolData(testExp) <- protocolData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
pData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
pData(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
pData(testExp) <- pData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
pData(testExp) <- pData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
pData(testExp) <- pData(testExp)[, 1:2]

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
varLabels(testExp)

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
varLabels(colData(testExp))

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
varLabels(testExp) <- paste0(varLabels(testExp), "a")

datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
varLabels(colData(testExp)[1:2]) <- 
    paste0(varLabels(colData(testExp)[1:2]), "a")

}
