% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{sampleNames,NanoStringExperiment-method}
\alias{sampleNames,NanoStringExperiment-method}
\alias{sampleNames<-,NanoStringExperiment,ANY-method}
\title{Access sample identifiers}
\usage{
\S4method{sampleNames}{NanoStringExperiment}(object)

\S4method{sampleNames}{NanoStringExperiment,ANY}(object) <- value
}
\arguments{
\item{object}{NanoStringExperiment object}

\item{value}{list sample names}
}
\value{
list of string sample (column) identifiers

NanoStringExperiment object
}
\description{
Shows variables used for rownames in colData.

Replace variables used for rownames in colData.
}
\examples{
data(exampleNSEData)
sampleNames(testExp)

data(exampleNSEData)
colData(testExp)[["testNames"]] <- paste0(dimnames(testExp)[[2]], "a")
sampleNames(testExp) <- colData(testExp)[["testNames"]]

}
