% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{signatures}
\alias{signatures}
\alias{signatures<-}
\alias{signatures<-,NanoStringExperiment,SignatureSet-method}
\title{Access signatures slot}
\usage{
signatures(object)

signatures(object) <- value

\S4method{signatures}{NanoStringExperiment,SignatureSet}(object) <- value
}
\value{
SignatureSet object

NanoStringExperiment object

NanoStringExperiment object
}
\description{
Access signatures slot to get \code{SignatureSet} object

Replace signatures slot with updated \code{SignatureSet}

Replace signatures slot with updated \code{SignatureSet}
}
\examples{
data(exampleNSEData)
signatures(testExp)

data(exampleNSEData)
signatures(testExp) <- NanoStringNCTools::SignatureSet()

data(exampleNSEData)
signatures(testExp) <- NanoStringNCTools::SignatureSet()

}
