% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-summary.R
\name{summary,NanoStringExperiment-method}
\alias{summary,NanoStringExperiment-method}
\title{Summarize assay data}
\usage{
\S4method{summary}{NanoStringExperiment}(
  object,
  MARGIN = 2L,
  GROUP = NULL,
  log2scale = TRUE,
  elt = "exprs",
  signatureScores = FALSE,
  ...
)
}
\description{
Get chosen summary statistics for a selected assay
}
\examples{
datadir <- system.file("data", package="NanoStringExperiment")
testExp <- readRDS(file.path(datadir, "testExp.rds"))
summary(testExp)

}
