% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-accessors.R
\name{svarLabels,NanoStringExperiment-method}
\alias{svarLabels,NanoStringExperiment-method}
\title{Access sample metadata varible names}
\usage{
\S4method{svarLabels}{NanoStringExperiment}(object)
}
\description{
It is recommended to use the SummarizedExperiment method colData instead.
This is a convenience method for backwards compatibility.
}
