% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringExperiment-summary.R
\name{with,NanoStringExperiment-method}
\alias{with,NanoStringExperiment-method}
\title{Evaluate expression on NanoStringExperiment list}
\usage{
\S4method{with}{NanoStringExperiment}(data, expr, ...)
}
\description{
Performs expression within NanoStringExperiment list
}
